Update Auto Setup (#19)
* Refactor auto initialization and holo params

* tmp

* replace old practice tuner constants with updated file

* fix errors when running code

* various changes to make pathplanner work

* add a new auto

---------
