+++ src/main/deploy/pathplanner/autos/Simple Coral Auto.auto
@@ -1,0 +1,67 @@
+{
+  "version": "2025.0",
+  "command": {
+    "type": "sequential",
+    "data": {
+      "commands": [
+        {
+          "type": "wait",
+          "data": {
+            "waitTime": 0.25
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Top Intake to Reef TR"
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Reef TR to Top Intake"
+          }
+        },
+        {
+          "type": "wait",
+          "data": {
+            "waitTime": 0.25
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Top Intake to Reef TL"
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Reef TL to Top Intake"
+          }
+        },
+        {
+          "type": "wait",
+          "data": {
+            "waitTime": 0.25
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Top Intake to Reef L"
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Reef L to Top Intake"
+          }
+        }
+      ]
+    }
+  },
+  "resetOdom": true,
+  "folder": null,
+  "choreoAuto": false
+}
