+++ src/main/deploy/pathplanner/autos/Test Auto.auto
@@ -1,0 +1,37 @@
+{
+  "version": "2025.0",
+  "command": {
+    "type": "sequential",
+    "data": {
+      "commands": [
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Test Forward"
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Test Backward"
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Test Turn"
+          }
+        },
+        {
+          "type": "path",
+          "data": {
+            "pathName": "Test Circle"
+          }
+        }
+      ]
+    }
+  },
+  "resetOdom": true,
+  "folder": "Test Autos",
+  "choreoAuto": false
+}
