+++ src/main/deploy/pathplanner/paths/Top Intake to Reef TL.path
@@ -3,29 +3,29 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 2.0,
-        "y": 7.0
+        "x": 0.9520512820485846,
+        "y": 6.759935897440073
       },
       "prevControl": null,
       "nextControl": {
-        "x": 3.0,
-        "y": 7.0
+        "x": 1.9737820512820505,
+        "y": 6.625769230769265
       },
       "isLocked": false,
-      "linkedName": null
+      "linkedName": "Top Intake"
     },
     {
       "anchor": {
-        "x": 4.0,
-        "y": 7.0
+        "x": 3.7798717948717937,
+        "y": 5.284102564102564
       },
       "prevControl": {
-        "x": 3.0,
-        "y": 7.0
+        "x": 2.8613461538461538,
+        "y": 5.604038461538462
       },
       "nextControl": null,
       "isLocked": false,
-      "linkedName": null
+      "linkedName": "Reef TL"
     }
   ],
   "rotationTargets": [],
@@ -42,13 +42,13 @@
   },
   "goalEndState": {
     "velocity": 0,
-    "rotation": 0.0
+    "rotation": -59.99999999999999
   },
   "reversed": false,
-  "folder": "Test Paths",
+  "folder": null,
   "idealStartingState": {
     "velocity": 0,
-    "rotation": 0.0
+    "rotation": 125.00000000000001
   },
   "useDefaultConstraints": true
 }
