+++ src/main/deploy/pathplanner/paths/Top Intake to Reef TR.path
@@ -3,29 +3,29 @@
   "waypoints": [
     {
       "anchor": {
-        "x": 2.0,
-        "y": 7.0
+        "x": 0.9520512820485846,
+        "y": 6.759935897440073
       },
       "prevControl": null,
       "nextControl": {
-        "x": 3.0,
-        "y": 7.0
+        "x": 2.4072435897435893,
+        "y": 7.306923076923113
       },
       "isLocked": false,
-      "linkedName": null
+      "linkedName": "Top Intake"
     },
     {
       "anchor": {
-        "x": 4.0,
-        "y": 7.0
+        "x": 5.214423076923076,
+        "y": 5.2221794871794875
       },
       "prevControl": {
-        "x": 3.0,
-        "y": 7.0
+        "x": 4.512628205128205,
+        "y": 6.708333333333334
       },
       "nextControl": null,
       "isLocked": false,
-      "linkedName": null
+      "linkedName": "Reef TR"
     }
   ],
   "rotationTargets": [],
@@ -42,13 +42,13 @@
   },
   "goalEndState": {
     "velocity": 0,
-    "rotation": 0.0
+    "rotation": -119.99999999999999
   },
   "reversed": false,
-  "folder": "Test Paths",
+  "folder": null,
   "idealStartingState": {
     "velocity": 0,
-    "rotation": 0.0
+    "rotation": 125.00000000000001
   },
   "useDefaultConstraints": true
 }
