+++ src/main/java/frc/robot/subsystems/CommandSwerveDrivetrain.java
@@ -10,9 +10,6 @@ import com.ctre.phoenix6.swerve.SwerveDrivetrainConstants;
 import com.ctre.phoenix6.swerve.SwerveModuleConstants;
 import com.ctre.phoenix6.swerve.SwerveRequest;
 import com.pathplanner.lib.auto.AutoBuilder;
-import com.pathplanner.lib.config.PIDConstants;
-import com.pathplanner.lib.controllers.PPHolonomicDriveController;
-import com.pathplanner.lib.controllers.PathFollowingController;
 import com.pathplanner.lib.util.DriveFeedforwards;
 
 import edu.wpi.first.math.Matrix;
@@ -28,16 +25,16 @@ import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj.Notifier;
 import edu.wpi.first.wpilibj.RobotController;
 import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.Subsystem;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.robot.Config;
-import frc.robot.generated.CompBotTunerConstants.TunerSwerveDrivetrain;
-import frc.robot.systems.autonomous.PathPlannableSubsystem;
+import frc.robot.generated.PracticeBotTunerConstants.TunerSwerveDrivetrain;
 
 /**
  * Class that extends the Phoenix 6 SwerveDrivetrain class and implements
  * Subsystem so it can easily be used in command-based projects.
  */
-public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements PathPlannableSubsystem {
+public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements Subsystem {
     private static final double kSimLoopPeriod = 0.005; // 5 ms
     private Notifier m_simNotifier = null;
     private double m_lastSimTime;
@@ -294,9 +291,6 @@ public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements Pa
         return this.getState().Pose;
     }
 
-    // The SwerveDriveTrain class already has a resetPose() function.
-
-    @Override
     public ChassisSpeeds getRobotRelativeChassisSpeeds() {
         return this.getState().Speeds;
     }
@@ -309,7 +303,6 @@ public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements Pa
      * we will need to reverse the feedforwards for modules that have been flipped.
      * See the output argument in {@link AutoBuilder#configure}
      */
-    @Override
     public void driveRobotRelative(ChassisSpeeds robotRelativeSpeeds, DriveFeedforwards driveFeedforwards) {
         ChassisSpeeds targetSpeeds = ChassisSpeeds.discretize(robotRelativeSpeeds, this.discretizationDelta);
         this.setControl(new SwerveRequest.ApplyRobotSpeeds()
@@ -318,22 +311,5 @@ public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements Pa
             .withWheelForceFeedforwardsY(driveFeedforwards.robotRelativeForcesY())
             .withDesaturateWheelSpeeds(true)
         );
-    }
-
-    @Override
-    public PathFollowingController getPathFollowingController() {
-        return new PPHolonomicDriveController(
-            new PIDConstants(5,0,0),
-            new PIDConstants(5,0,0)
-        );
-    }
-
-    @Override
-    public boolean isRedAlliance() {
-        var alliance = DriverStation.getAlliance();
-        if (alliance.isPresent()) {
-            return alliance.get() == DriverStation.Alliance.Red;
-        }
-        return false;
     }
 }
