+++ src/main/java/frc/robot/systems/autonomous/AutonomousControllerImpl.java
@@ -5,9 +5,13 @@ import java.util.function.BooleanSupplier;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.commands.PathPlannerAuto;
+import com.pathplanner.lib.config.PIDConstants;
+import com.pathplanner.lib.controllers.PPHolonomicDriveController;
+import com.pathplanner.lib.controllers.PathFollowingController;
 
 import edu.wpi.first.wpilibj2.command.Commands;
 import frc.robot.Config;
+import frc.robot.subsystems.CommandSwerveDrivetrain;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
@@ -25,22 +29,23 @@ public class AutonomousControllerImpl implements AutonomousController {
 
     private final SendableChooser<Command> autoChooser;
 
-    private AutonomousControllerImpl(Config config, PathPlannableSubsystem driveSystem) {
+    private AutonomousControllerImpl(Config config, CommandSwerveDrivetrain driveSystem) {
         // Boolean supplier that controls when the path will be mirrored for the red
         // alliance
         // This will flip the path being followed to the red side of the field.
         // THE ORIGIN WILL REMAIN ON THE BLUE SIDE
-        BooleanSupplier requiresFlip = driveSystem::isRedAlliance;
+        BooleanSupplier requiresFlip = Config::isRedAlliance;
 
         AutoBuilder.configure(
             driveSystem::getPose,
             driveSystem::resetPose,
             driveSystem::getRobotRelativeChassisSpeeds,
             driveSystem::driveRobotRelative,
-            driveSystem.getPathFollowingController(),
+            createPathFollowingController(),
             config.generatedConfig,
             requiresFlip,
-            driveSystem);
+            driveSystem
+        );
         loadAutos();
 
         autoChooser = AutoBuilder.buildAutoChooser();
@@ -49,11 +54,13 @@ public class AutonomousControllerImpl implements AutonomousController {
 
     private void loadAutos() {
         autos = Map.of(
-            "Config", new PathPlannerAuto("Config")
+            "Config", new PathPlannerAuto("Config"),
+            "Test Auto", new PathPlannerAuto("Test Auto"),
+            "Simple Coral Auto", new PathPlannerAuto("Simple Coral Auto")
         );
     }
 
-    public static synchronized AutonomousControllerImpl initialize(Config config, PathPlannableSubsystem driveSystem) {
+    public static synchronized AutonomousControllerImpl initialize(Config config, CommandSwerveDrivetrain driveSystem) {
         controller = new AutonomousControllerImpl(config, driveSystem);
         return controller;
     }
@@ -68,7 +75,7 @@ public class AutonomousControllerImpl implements AutonomousController {
     @Override
     public void runInit() {
         CommandScheduler.getInstance().cancelAll();
-        autos.get("Config").schedule();
+        autos.get("Test Auto").schedule();
     }
 
     @Override
@@ -77,5 +84,12 @@ public class AutonomousControllerImpl implements AutonomousController {
     @Override
     public void runExit() {
         Commands.print("No autonomous exit configured").schedule();
+    }
+
+    private PathFollowingController createPathFollowingController() {
+        return new PPHolonomicDriveController(
+            new PIDConstants(5, 0, 0),
+            new PIDConstants(5, 0, 0)
+        );
     }
 }
